requirements.txt
pytest~=6.2.3
eth-brownie>=1.14.6
rich~=10.2.0
